read test config from parameter server
